+++ ui/components/providers/workflow/forms/select-credentials-type/aws/credentials-type/aws-role-credentials-form.tsx
@@ -1,12 +1,11 @@
 import { Divider, Select, SelectItem, Spacer } from "@nextui-org/react";
 import { Control, UseFormSetValue, useWatch } from "react-hook-form";
 
+import { CredentialsRoleHelper } from "@/components/providers/workflow";
 import { CustomInput } from "@/components/ui/custom";
 import { AWSCredentialsRole } from "@/types";
 
-import { CredentialsRoleHelper } from "../../credentials-role-helper";
-
-export const AWSCredentialsRoleForm = ({
+export const AWSRoleCredentialsForm = ({
   control,
   setValue,
   externalId,