+++ ui/components/providers/workflow/forms/select-credentials-type/aws/credentials-type/index.ts
@@ -1,0 +1,2 @@
+export * from "./aws-role-credentials-form";
+export * from "./aws-static-credentials-form";